backend/requirements.txt
# FastAPI
fastapi>=0.115.6
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20

# Database
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.1

# Pydantic
pydantic>=2.10.4
pydantic-settings>=2.7.1

# Qdrant
qdrant-client>=1.12.1

# Embedding (bge-m3)
sentence-transformers>=4.0

# OCR (surya brings pillow + pypdfium2)
surya-ocr>=0.17.1

# PDF extraction
pdfplumber>=0.11.4

# LLM (OpenRouter)
httpx>=0.28.1

# Utilities
python-dotenv>=1.0.1
